+++ src/main/java/tk/parking/app/repo/ParkingRepo.java
@@ -3,5 +3,5 @@ package tk.parking.app.repo;
 import org.springframework.data.repository.CrudRepository;
 import tk.parking.app.entity.Parking;
 
-public interface ParkingRepo extends CrudRepository<Parking, Long> {
+public interface ParkingRepo extends CrudRepository<Parking, Integer> {
 }
